+++ client/views/common/footer.js
@@ -1,10 +1,7 @@
 Template.footer.helpers({
   footerCode: function(){
     return getSetting('footerCode');
-  },
-  extraCode: function(){
-    return getSetting('extraCode');
-  },  
+  }, 
   distanceFromTop: function(){
     var distanceFromTop = parseInt(Session.get('distanceFromTop'))+70+20;
     if(Meteor.Router.page()!='posts_digest' && !Session.get('allPostsLoaded'))