+++ common/comment.js
@@ -44,22 +44,26 @@ Meteor.methods({
     if(!this.isSimulation){
       if(parentCommentId){
         // child comment
-        var parentComment=Comments.find(parentCommentId);
-        var parentUser=Meteor.users.find(parentComment.userId);
+        var parentComment=Comments.findOne(parentCommentId);
+        var parentUser=Meteor.users.findOne(parentComment.userId);
 
         properties.parentCommentId = parentCommentId;
         properties.parentAuthorId = parentComment.userId;
         properties.parentAuthorName = getDisplayName(parentUser);
 
-        createNotification('newReply', properties, parentUser, user);
+        // do not notify users of their own actions (i.e. they're replying to themselves)
+        if(parentUser._id != user._id)
+          Meteor.call('createNotification','newReply', properties, parentUser, user);
 
-        if(parentComment.userId!=post.userId){
-          // if the original poster is different from the author of the parent comment, notify them too
-          createNotification('newComment', properties, postUser, Meteor.user());
-        }
+        // if the original poster is different from the author of the parent comment, notify them too
+        if(postUser._id != user._id && parentComment.userId != post.userId)
+          Meteor.call('createNotification','newComment', properties, postUser, user);
+
       }else{
         // root comment
-        createNotification('newComment', properties, postUser, Meteor.user());
+        // don't notify users of their own comments
+        if(postUser._id != user._id)
+          Meteor.call('createNotification','newComment', properties, postUser, Meteor.user());
       }
     }
     return properties;